Added the Minsk project
The CodeAnalysis files have been migrated out of mc.csproj to
Minsk.csproj. To facilitate interaction with the Binder and Evaluator
that are internal classes to Minsk, from mc, we've introduced new
classes, EvaluationResult and Compilation, that abstract away the
internals.
