+++ .vscode/cspell.json
@@ -1,0 +1,18 @@
+// cSpell Settings
+{
+    // Version of the setting file.  Always 0.2
+    "version": "0.2",
+    // language - current active spelling language
+    "language": "en-GB",
+    // words - list of words to be always considered correct
+    "words": [
+        "lexer"
+    ],
+    // Specify file types to spell check.
+    "enabledLanguageIds": [
+        "csharp",
+        "markdown",
+        "plaintext",
+        "yml"
+    ]
+}
