+++ Minsk/CodeAnalysis/Binding/Binder.cs
@@ -1,0 +1,58 @@
+using System.Collections.Immutable;
+using Minsk.CodeAnalysis.Syntax;
+
+namespace Minsk.CodeAnalysis.Binding;
+
+internal sealed class Binder
+{
+    private readonly List<string> _diagnostics = new();
+
+    public BoundExpression BindExpression(ExpressionSyntax syntax)
+    {
+        return syntax.Kind switch
+        {
+            SyntaxKind.LiteralExpression => BindLiteralExpression((LiteralExpressionSyntax)syntax),
+            SyntaxKind.BinaryExpression => BindBinaryExpression((BinaryExpressionSyntax)syntax),
+            SyntaxKind.UnaryExpression => BindUnaryExpression((UnaryExpressionSyntax)syntax),
+            SyntaxKind.ParenthesizedExpression => BindExpression(((ParenthesizedExpressionSyntax)syntax).Expression),
+            _ => throw new Exception($"Unexpected syntax {syntax.Kind}"),
+        };
+    }
+
+    public IReadOnlyCollection<string> Diagnostics => _diagnostics.ToImmutableList();
+
+    private BoundExpression BindLiteralExpression(LiteralExpressionSyntax syntax)
+    {
+        var value = syntax.Value ?? 0;
+        return new BoundLiteralExpression(value);
+    }
+
+    private BoundExpression BindBinaryExpression(BinaryExpressionSyntax syntax)
+    {
+        var boundLeft = BindExpression(syntax.Left);
+        var boundRight = BindExpression(syntax.Right);
+        var boundOperator = BoundBinaryOperator.Bind(syntax.OperatorToken.Kind, boundLeft.Type, boundRight.Type);
+
+        if (boundOperator == null)
+        {
+            _diagnostics.Add($"Binary operator '{syntax.OperatorToken.Text}' is not defined for types {boundLeft.Type} and {boundRight.Type}.");
+            return boundLeft;
+        }
+
+        return new BoundBinaryExpression(boundLeft, boundOperator, boundRight);
+    }
+
+    private BoundExpression BindUnaryExpression(UnaryExpressionSyntax syntax)
+    {
+        var boundOperand = BindExpression(syntax.Operand);
+        var boundOperator = BoundUnaryOperator.Bind(syntax.OperatorToken.Kind, boundOperand.Type);
+
+        if (boundOperator == null)
+        {
+            _diagnostics.Add($"Unary operator '{syntax.OperatorToken.Text}' is not defined for type {boundOperand.Type}.");
+            return boundOperand;
+        }
+
+        return new BoundUnaryExpression(boundOperator, boundOperand);
+    }
+}
