+++ Minsk/CodeAnalysis/Binding/BoundBinaryOperatorKind.cs
@@ -1,0 +1,13 @@
+namespace Minsk.CodeAnalysis.Binding;
+
+internal enum BoundBinaryOperatorKind
+{
+    Addition,
+    Subtraction,
+    Multiplication,
+    Division,
+    LogicalAnd,
+    LogicalOr,
+    Equals,
+    NotEquals
+}
