+++ Minsk/CodeAnalysis/Binding/BoundExpression.cs
@@ -1,0 +1,6 @@
+namespace Minsk.CodeAnalysis.Binding;
+
+internal abstract class BoundExpression : BoundNode
+{
+    public abstract Type Type { get; }
+}
