+++ Minsk/CodeAnalysis/Binding/BoundNode.cs
@@ -1,0 +1,6 @@
+namespace Minsk.CodeAnalysis.Binding;
+
+internal abstract class BoundNode
+{
+    public abstract BoundNodeKind Kind { get; }
+}
