+++ Minsk/CodeAnalysis/Binding/BoundNodeKind.cs
@@ -1,0 +1,8 @@
+namespace Minsk.CodeAnalysis.Binding;
+
+internal enum BoundNodeKind
+{
+    UnaryExpression,
+    LiteralExpression,
+    BinaryExpression
+}
