+++ Minsk/CodeAnalysis/Binding/BoundUnaryOperatorKind.cs
@@ -1,0 +1,8 @@
+namespace Minsk.CodeAnalysis.Binding;
+
+internal enum BoundUnaryOperatorKind
+{
+    Identity,
+    Negation,
+    LogicalNegation
+}
