+++ Minsk/CodeAnalysis/Syntax/BinaryExpressionSyntax.cs
@@ -1,0 +1,23 @@
+namespace Minsk.CodeAnalysis.Syntax;
+
+public sealed class BinaryExpressionSyntax : ExpressionSyntax
+{
+    public BinaryExpressionSyntax(ExpressionSyntax left, SyntaxToken operatorToken, ExpressionSyntax right)
+    {
+        Left = left;
+        OperatorToken = operatorToken;
+        Right = right;
+    }
+
+    public override SyntaxKind Kind => SyntaxKind.BinaryExpression;
+    public ExpressionSyntax Left { get; }
+    public SyntaxToken OperatorToken { get; }
+    public ExpressionSyntax Right { get; }
+
+    public override IEnumerable<SyntaxNode> GetChildren()
+    {
+        yield return Left;
+        yield return OperatorToken;
+        yield return Right;
+    }
+}
