+++ Minsk/CodeAnalysis/Syntax/ExpressionSyntax.cs
@@ -1,0 +1,3 @@
+namespace Minsk.CodeAnalysis.Syntax;
+
+public abstract class ExpressionSyntax : SyntaxNode { }
