+++ Minsk/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,0 +1,32 @@
+namespace Minsk.CodeAnalysis.Syntax;
+
+public enum SyntaxKind
+{
+    // Tokens
+    BadToken,
+    EndOfFileToken,
+    WhitespaceToken,
+    NumberToken,
+    PlusToken,
+    MinusToken,
+    StarToken,
+    SlashToken,
+    BangToken,
+    AmpersandAmpersandToken,
+    PipePipeToken,
+    EqualsEqualsToken,
+    BangEqualsToken,
+    OpenParenthesisToken,
+    CloseParenthesisToken,
+    IdentifierToken,
+
+    // Keywords
+    FalseKeyword,
+    TrueKeyword,
+
+    // Expressions
+    LiteralExpression,
+    UnaryExpression,
+    BinaryExpression,
+    ParenthesizedExpression
+}
