+++ Minsk/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,0 +1,7 @@
+namespace Minsk.CodeAnalysis.Syntax;
+
+public abstract class SyntaxNode
+{
+    public abstract SyntaxKind Kind { get; }
+    public abstract IEnumerable<SyntaxNode> GetChildren();
+}
