+++ Minsk/CodeAnalysis/Syntax/SyntaxTree.cs
@@ -1,0 +1,23 @@
+using System.Collections.Immutable;
+
+namespace Minsk.CodeAnalysis.Syntax;
+
+public sealed class SyntaxTree
+{
+    public SyntaxTree(IEnumerable<string> diagnostics, ExpressionSyntax root, SyntaxToken endOfFileToken)
+    {
+        Diagnostics = diagnostics.ToImmutableArray();
+        Root = root;
+        EndOfFileToken = endOfFileToken;
+    }
+
+    public IReadOnlyList<string> Diagnostics { get; }
+    public ExpressionSyntax Root { get; }
+    public SyntaxToken EndOfFileToken { get; }
+
+    public static SyntaxTree Parse(string text)
+    {
+        var parser = new Parser(text);
+        return parser.Parse();
+    }
+}
