--- mc/CodeAnalysis/Binding/BoundBinaryOperatorKind.cs
@@ -1,14 +1,0 @@
-namespace Minsk.CodeAnalysis.Binding
-{
-    internal enum BoundBinaryOperatorKind
-    {
-        Addition,
-        Subtraction,
-        Multiplication,
-        Division,
-        LogicalAnd,
-        LogicalOr,
-        Equals,
-        NotEquals
-    }
-}
