--- mc/CodeAnalysis/Binding/BoundExpression.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Minsk.CodeAnalysis.Binding
-{
-    internal abstract class BoundExpression : BoundNode
-    {
-        public abstract Type Type { get; }
-    }
-}
