--- mc/CodeAnalysis/Binding/BoundLiteralExpression.cs
@@ -1,16 +1,0 @@
-using System;
-
-namespace Minsk.CodeAnalysis.Binding
-{
-    internal sealed class BoundLiteralExpression : BoundExpression
-    {
-        public BoundLiteralExpression(object value)
-        {
-            Value = value;
-        }
-
-        public override BoundNodeKind Kind => BoundNodeKind.LiteralExpression;
-        public override Type Type => Value.GetType();
-        public object Value { get; }
-    }
-}
