--- mc/CodeAnalysis/Binding/BoundNodeKind.cs
@@ -1,9 +1,0 @@
-namespace Minsk.CodeAnalysis.Binding
-{
-    internal enum BoundNodeKind
-    {
-        UnaryExpression,
-        LiteralExpression,
-        BinaryExpression
-    }
-}
