--- mc/CodeAnalysis/Syntax/ExpressionSyntax.cs
@@ -1,4 +1,0 @@
-namespace Minsk.CodeAnalysis.Syntax
-{
-    public abstract class ExpressionSyntax : SyntaxNode { }
-}
