--- mc/CodeAnalysis/Syntax/Parser.cs
@@ -1,132 +1,0 @@
-using System.Collections.Generic;
-using System.Collections.Immutable;
-
-namespace Minsk.CodeAnalysis.Syntax
-{
-    internal sealed class Parser
-    {
-        private readonly SyntaxToken[] _tokens;
-        private int _position;
-        private readonly List<string> _diagnostics = new();
-
-        public Parser(string text)
-        {
-            var tokens = new List<SyntaxToken>();
-
-            var lexer = new Lexer(text);
-            SyntaxToken token;
-            do
-            {
-                token = lexer.Lex();
-
-                if (token.Kind != SyntaxKind.BadToken &&
-                    token.Kind != SyntaxKind.WhitespaceToken)
-                {
-                    tokens.Add(token);
-                }
-            }
-            while (token.Kind != SyntaxKind.EndOfFileToken);
-
-            _tokens = tokens.ToArray();
-            _diagnostics.AddRange(lexer.Diagnostics);
-        }
-
-        public IEnumerable<string> Diagnostics
-        {
-            get
-            {
-                return _diagnostics.ToImmutableArray();
-            }
-        }
-
-        private SyntaxToken Current => Peek(0);
-
-        public SyntaxTree Parse()
-        {
-            var expression = ParseExpression();
-            var eofToken = MatchToken(SyntaxKind.EndOfFileToken);
-            return new SyntaxTree(Diagnostics, expression, eofToken);
-        }
-
-        private ExpressionSyntax ParseExpression(int parentPrecedence = 0)
-        {
-            ExpressionSyntax left;
-            var unaryOperatorPrecedence = Current.Kind.GetUnaryOperatorPrecedence();
-            if (unaryOperatorPrecedence != 0 && unaryOperatorPrecedence >= parentPrecedence)
-            {
-                var operatorToken = NextToken();
-                var operand = ParseExpression(unaryOperatorPrecedence);
-                left = new UnaryExpressionSyntax(operatorToken, operand);
-            }
-            else
-            {
-                left = ParsePrimaryExpression();
-            }
-
-            while (true)
-            {
-                var precedence = Current.Kind.GetBinaryOperatorPrecedence();
-                if (precedence == 0 || precedence <= parentPrecedence)
-                {
-                    break;
-                }
-
-                var operatorToken = NextToken();
-                var right = ParseExpression(precedence);
-                left = new BinaryExpressionSyntax(left, operatorToken, right);
-            }
-
-            return left;
-        }
-
-        private ExpressionSyntax ParsePrimaryExpression()
-        {
-            switch (Current.Kind)
-            {
-                case SyntaxKind.OpenParenthesisToken:
-                    var left = NextToken();
-                    var expression = ParseExpression();
-                    var right = MatchToken(SyntaxKind.CloseParenthesisToken);
-                    return new ParenthesizedExpressionSyntax(left, expression, right);
-
-                case SyntaxKind.FalseKeyword:
-                case SyntaxKind.TrueKeyword:
-                    var keywordToken = NextToken();
-                    var value = keywordToken.Kind == SyntaxKind.TrueKeyword;
-                    return new LiteralExpressionSyntax(keywordToken, value);
-
-                default:
-                    var numberToken = MatchToken(SyntaxKind.NumberToken);
-                    return new LiteralExpressionSyntax(numberToken);
-            }
-        }
-
-        private SyntaxToken MatchToken(SyntaxKind kind)
-        {
-            if (Current.Kind == kind)
-            {
-                return NextToken();
-            }
-
-            _diagnostics.Add($"ERROR: Unexpected token <{Current.Kind}>, expected <{kind}>");
-            return new SyntaxToken(kind, Current.Position, null, null);
-        }
-
-        private SyntaxToken NextToken()
-        {
-            var current = Current;
-            _position++;
-            return current;
-        }
-
-        private SyntaxToken Peek(int offset)
-        {
-            var index = _position + offset;
-            if (index >= _tokens.Length)
-            {
-                return _tokens[^1];
-            }
-            return _tokens[index];
-        }
-    }
-}
