--- mc/CodeAnalysis/Syntax/SyntaxFacts.cs
@@ -1,39 +1,0 @@
-using System;
-
-namespace Minsk.CodeAnalysis.Syntax
-{
-    internal static class SyntaxFacts
-    {
-        public static int GetUnaryOperatorPrecedence(this SyntaxKind kind)
-        {
-            return kind switch
-            {
-                SyntaxKind.PlusToken or SyntaxKind.MinusToken or SyntaxKind.BangToken => 6,
-                _ => 0,
-            };
-        }
-
-        public static int GetBinaryOperatorPrecedence(this SyntaxKind kind)
-        {
-            return kind switch
-            {
-                SyntaxKind.StarToken or SyntaxKind.SlashToken => 5,
-                SyntaxKind.PlusToken or SyntaxKind.MinusToken => 4,
-                SyntaxKind.EqualsEqualsToken or SyntaxKind.BangEqualsToken => 3,
-                SyntaxKind.AmpersandAmpersandToken => 2,
-                SyntaxKind.PipePipeToken => 1,
-                _ => 0,
-            };
-        }
-
-        public static SyntaxKind GetKeywordKind(string text)
-        {
-            return text switch
-            {
-                "true" => SyntaxKind.TrueKeyword,
-                "false" => SyntaxKind.FalseKeyword,
-                _ => SyntaxKind.IdentifierToken,
-            };
-        }
-    }
-}
