--- mc/CodeAnalysis/Syntax/SyntaxKind.cs
@@ -1,33 +1,0 @@
-namespace Minsk.CodeAnalysis.Syntax
-{
-    public enum SyntaxKind
-    {
-        // Tokens
-        BadToken,
-        EndOfFileToken,
-        WhitespaceToken,
-        NumberToken,
-        PlusToken,
-        MinusToken,
-        StarToken,
-        SlashToken,
-        BangToken,
-        AmpersandAmpersandToken,
-        PipePipeToken,
-        EqualsEqualsToken,
-        BangEqualsToken,
-        OpenParenthesisToken,
-        CloseParenthesisToken,
-        IdentifierToken,
-
-        // Keywords
-        FalseKeyword,
-        TrueKeyword,
-
-        // Expressions
-        LiteralExpression,
-        UnaryExpression,
-        BinaryExpression,
-        ParenthesizedExpression
-    }
-}
