--- mc/CodeAnalysis/Syntax/SyntaxNode.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-
-namespace Minsk.CodeAnalysis.Syntax
-{
-    public abstract class SyntaxNode
-    {
-        public abstract SyntaxKind Kind { get; }
-        public abstract IEnumerable<SyntaxNode> GetChildren();
-    }
-}
