+++ minsk.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 16.6.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "mc", "mc\mc.csproj", "{900F24CA-2D81-4A79-BEE8-ABAC91804F7E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Minsk", "Minsk\Minsk.csproj", "{67AE4A56-C700-4705-98AD-57135C51D96A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,17 @@ Global
 		{900F24CA-2D81-4A79-BEE8-ABAC91804F7E}.Release|x64.Build.0 = Release|Any CPU
 		{900F24CA-2D81-4A79-BEE8-ABAC91804F7E}.Release|x86.ActiveCfg = Release|Any CPU
 		{900F24CA-2D81-4A79-BEE8-ABAC91804F7E}.Release|x86.Build.0 = Release|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Debug|x64.Build.0 = Debug|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Debug|x86.Build.0 = Debug|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Release|x64.ActiveCfg = Release|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Release|x64.Build.0 = Release|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Release|x86.ActiveCfg = Release|Any CPU
+		{67AE4A56-C700-4705-98AD-57135C51D96A}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
